// src/models/Word.ts
export class Word {

    id: number;
    num: number;
    lecId: number;

    col01?: string;
    col02?: string;
    col03?: string;
    col04?: string;
    col05?: string;
    col06?: string;
    col07?: string;
    col08?: string;
    col09?: string;
    col10?: string;
    col11?: string;
    col12?: string;
    col13?: string;
    col14?: string;
    col15?: string;
    col16?: string;

}